+++ app/src/main/java/com/davay/android/core/data/network/model/getmatches/GetSessionRequest.kt
@@ -1,4 +1,4 @@
-package com.davay.android.feature.coincidences.data.network.models
+package com.davay.android.core.data.network.model.getmatches
 
 class GetSessionRequest(val sessionId: String, val userId: String) {
     val path: String = "api/sessions/$sessionId/get_matched_movies/"
