+++ app/src/main/java/com/davay/android/core/domain/api/GetMatchesRepository.kt
@@ -1,10 +1,10 @@
-package com.davay.android.feature.coincidences.domain.api
+package com.davay.android.core.domain.api
 
 import com.davay.android.core.domain.models.ErrorType
 import com.davay.android.core.domain.models.MovieDetails
 import com.davay.android.core.domain.models.Result
 import kotlinx.coroutines.flow.Flow
 
-interface CoincidencesRepository {
+interface GetMatchesRepository {
     fun getMatches(sessionId: String): Flow<Result<List<MovieDetails>, ErrorType>>
 }
