+++ app/src/main/java/com/davay/android/feature/coincidences/di/CoincidencesDomainModule.kt
@@ -1,11 +1,11 @@
 package com.davay.android.feature.coincidences.di
 
+import com.davay.android.core.domain.api.GetMatchesRepository
 import com.davay.android.core.domain.impl.CommonWebsocketInteractor
+import com.davay.android.core.domain.impl.GetMatchesUseCase
 import com.davay.android.core.domain.lounchcontrol.api.FirstTimeFlagRepository
 import com.davay.android.feature.coincidences.domain.CoincidencesInteractorImpl
-import com.davay.android.feature.coincidences.domain.GetMatchesUseCase
 import com.davay.android.feature.coincidences.domain.api.CoincidencesInteractor
-import com.davay.android.feature.coincidences.domain.api.CoincidencesRepository
 import dagger.Module
 import dagger.Provides
 
@@ -20,7 +20,7 @@ class CoincidencesDomainModule {
 
     @Provides
     fun provideGetMatchesUseCase(
-        repository: CoincidencesRepository,
+        repository: GetMatchesRepository,
         commonWebsocketInteractor: CommonWebsocketInteractor
     ) = GetMatchesUseCase(repository, commonWebsocketInteractor)
 }
